Registro de trabajos y refinacion de gastos

+++ database/migrations/2024_09_08_234157_create_empleado_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->float('cedula');
             $table->date('fecha_inicio');
             $table->string('foto');
-            $table->decimal('numero');
-            $table->decimal('numero_emergencia');
+            $table->float('numero');
+            $table->float('numero_emergencia');
             $table->enum('tipo_contrato',['Prestacion_de_servicios','Horas','Planta']);
             $table->enum('estado',['Activo','Inactivo','Enfermo']);
             $table->string('nss'); //Nro de Seguridad Social
